# Make the collector exporter mutate data, and remove unnecessary CopyTo