--- a/exporter/collector/internal/normalization/disabled_normalizer.go
+++ b/exporter/collector/internal/normalization/disabled_normalizer.go
@@ -33,7 +33,8 @@ func NewDisabledNormalizer() Normalizer {
 
 type disabledNormalizer struct{}
 
-// NormalizeExponentialHistogramDataPoint returns the point without normalizing.
+// NormalizeExponentialHistogramDataPoint ensures the start time is before the
+// end time, but does not normalize points.
 func (d *disabledNormalizer) NormalizeExponentialHistogramDataPoint(point pmetric.ExponentialHistogramDataPoint, _ uint64) bool {
 	if !point.StartTimestamp().AsTime().Before(point.Timestamp().AsTime()) {
 		// Handle explicit reset points.
@@ -43,7 +44,8 @@ func (d *disabledNormalizer) NormalizeExponentialHistogramDataPoint(point pmetri
 	return true
 }
 
-// NormalizeHistogramDataPoint returns the point without normalizing.
+// NormalizeHistogramDataPoint ensures the start time is before the
+// end time, but does not normalize points.
 func (d *disabledNormalizer) NormalizeHistogramDataPoint(point pmetric.HistogramDataPoint, _ uint64) bool {
 	if !point.StartTimestamp().AsTime().Before(point.Timestamp().AsTime()) {
 		// Handle explicit reset points.
@@ -53,7 +55,8 @@ func (d *disabledNormalizer) NormalizeHistogramDataPoint(point pmetric.Histogram
 	return true
 }
 
-// NormalizeNumberDataPoint returns the point without normalizing.
+// NormalizeNumberDataPoint ensures the start time is before the
+// end time, but does not normalize points.
 func (d *disabledNormalizer) NormalizeNumberDataPoint(point pmetric.NumberDataPoint, _ uint64) bool {
 	if !point.StartTimestamp().AsTime().Before(point.Timestamp().AsTime()) {
 		// Handle explicit reset points.
@@ -63,7 +66,8 @@ func (d *disabledNormalizer) NormalizeNumberDataPoint(point pmetric.NumberDataPo
 	return true
 }
 
-// NormalizeSummaryDataPoint returns the point without normalizing.
+// NormalizeSummaryDataPoint ensures the start time is before the
+// end time, but does not normalize points.
 func (d *disabledNormalizer) NormalizeSummaryDataPoint(point pmetric.SummaryDataPoint, _ uint64) bool {
 	if !point.StartTimestamp().AsTime().Before(point.Timestamp().AsTime()) {
 		// Handle explicit reset points.