Looking at https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/pull/890#issuecomment-2344718704

This change makes the collector exporter now mutate data.  We were doing multiple layers additional copying in our exporter to prevent mutating data, but it is likely better for us to just be a mutating exporter, and let the framework handle it.

This keeps the collector code clean, and should have better CPU/Memory usage (even after the additional copy) since we are copying fewer times.

Benchmarks (including those added in https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/pull/893):

```
goos: linux
goarch: amd64
pkg: github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/collector/internal/normalization
cpu: AMD EPYC 7B12
                                              │   main.txt    │             mutating.txt              │
                                              │    sec/op     │    sec/op     vs base                 │
NormalizeNumberDataPoint-2                      1845.0n ± ∞ ¹   532.6n ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeHistogramDataPoint-2                   2314.0n ± ∞ ¹   615.7n ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeExopnentialHistogramDataPoint-2        2071.0n ± ∞ ¹   673.7n ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeSummaryDataPoint-2                     1718.0n ± ∞ ¹   465.9n ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeNumberDataPoint-2                 2094.0n ± ∞ ¹   826.3n ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeHistogramDataPoint-2               2.565µ ± ∞ ¹   1.043µ ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeExponentialHistogramDataPoint-2   2148.0n ± ∞ ¹   869.4n ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeSummaryDataPoint-2                1775.0n ± ∞ ¹   836.2n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                          2.049µ         709.9n        -65.36%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                              │   main.txt   │             mutating.txt             │
                                              │     B/op     │    B/op      vs base                 │
NormalizeNumberDataPoint-2                      408.00 ± ∞ ¹   16.00 ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeHistogramDataPoint-2                   536.00 ± ∞ ¹   32.00 ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeExopnentialHistogramDataPoint-2        600.00 ± ∞ ¹   48.00 ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeSummaryDataPoint-2                     320.00 ± ∞ ¹   16.00 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeNumberDataPoint-2                  504.0 ± ∞ ¹   128.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeHistogramDataPoint-2               745.0 ± ∞ ¹   256.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeExponentialHistogramDataPoint-2    792.0 ± ∞ ¹   256.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeSummaryDataPoint-2                 432.0 ± ∞ ¹   128.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                          520.6         67.33        -87.07%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                              │   main.txt   │             mutating.txt             │
                                              │  allocs/op   │  allocs/op   vs base                 │
NormalizeNumberDataPoint-2                      10.000 ± ∞ ¹   2.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeHistogramDataPoint-2                   13.000 ± ∞ ¹   3.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeExopnentialHistogramDataPoint-2        14.000 ± ∞ ¹   4.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
NormalizeSummaryDataPoint-2                      8.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeNumberDataPoint-2                 12.000 ± ∞ ¹   4.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeHistogramDataPoint-2              17.000 ± ∞ ¹   7.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeExponentialHistogramDataPoint-2   14.000 ± ∞ ¹   4.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
ResetNormalizeSummaryDataPoint-2                11.000 ± ∞ ¹   4.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                          12.09         3.191        -73.61%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```