@dashpole 2024-09-12
depends on https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/pull/893