@aabmass 2024-09-12
It's weird the benchmarks now have `± ∞ ¹ ` vs https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/pull/890 which had meaningful bounds. Any idea what's going on?