@dashpole 2024-09-19
It is an option to exporterhelper: https://pkg.go.dev/go.opentelemetry.io/collector/exporter/exporterhelper#WithCapabilities, which we will need to set in our exporters